Add Object ID converter
When using MongoDB as a datastore with Flask, there's a good chance you'll need to put an Object ID in a URL. While you can use Flask's string router to loosely match it, the converter can't guarantee the value is actually an Object ID. This can lead to `TypeError` or `bson.error.InvalidId` exceptions being raised unexpectedly.

`flask_simon.ObjectIDConverter` registers itself as a converter to validate Object IDs in URLs. By raising `ValidationError` in the `to_python()` method, other converters are given the chance to match.

Closes #5