Mock connecting to the database
When testing that `objectid` has been registered as a converter, `Simon.init_app()` needs to be called. It tries to connect to the database.

In order to ensure that Flask-Simon's tests can be run without a database connection, `simon.connection.connect()` is being mocked in `test_objectidconverter()`. The test has nothing to do with the database connection, so it can safely be ignored.