Return a 400 for invalid Object IDs
As originally implemented, the `objectid` converter would raise `ValidationError` for values that could not be converted into Object IDs. This resulted in a 404 being returned to the client. While this provides consistency with the built-in converters, semantically it is a bit flawed. A 404 implies that while the resource could not be found, there may be a time when it is found.

The converter is changing to instead call `abort(400)`. This will cause a 400 to be returned to the client. This makes more sense because not only will the resource never be found, the problem is actually with the request since the value is not a valid Object ID.