Replace use of replicaSet with replica_set
Version 0.2 of Simon switched the preferred name of the replica set argument when calling the `connect()` method.  This updates the one reference to `connect()` to use `replica_set` instead of `replicaSet`. It also updates the tests and sets the minimum version of Simon to 0.2.

While this version of Simon includes a change that isn't backwards compatible, Simon isn't considered production-ready yet, so I feel comfortable making the change.